create login handler
